bs4 malformed xml parsing + xml namespace detection
