+++ requirements.txt
@@ -25,7 +25,7 @@ IP2Location==8.10.0
 androguard==3.4.0a1
 apkid==2.1.5
 quark-engine==23.7.1
-frida==15.2.2
+frida==16.1.3
 tldextract==3.4.4
 openstep-parser==1.5.4
 svgutils==0.3.4
